requirements.txt
latch
biopython
requests